apps/todops-cli/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "todops"
version = "0.1.1"
description = "A minimal CLI for todo platform operations"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "minio==7.2.12",
    "requests>=2.25.0",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=0.19.0",
    "slack-sdk==3.37.0"
]

[project.scripts]
todops = "todops.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "coverage>=7.13.0",
    "pytest-cov>=7.0.0",
    "selenium>=4.39.0",
    "webdriver-manager>=4.0.2"
]
